VIC-4717 Prevent eyes appearing behind time (#3931)
The show clock behavior demonstrates a corner case where we are
transitioning between on-screen eyes and non-eye display geometry,
like a clock face. In these cases there is no content in the
Procedural_Eyes layer since the eyes are just part of the png
sequence, but keepalive performances set up before this anim starts
are hanging around and inducing keepalive events resulting in eyes
displayed behind the intended content. At present, keepalives are only
updated if _streamingAnimation == nullptr anyway, so this change just
clears out any existing Keepalive keyframes in the
_proceduralTrackComponent any time we initialize a new anim for
streaming. If we eventually decide we want to have an animation run
with keepalive eyes in the background, this will need to be addressed
across the entire keepalive system.

* Make ProceduralClock use BaseStationTimer for internal transitions

BehaviorProceduralClock was using system time to determine how long
it should display the clock image for. Since its a delta, this has no
benefit I can think of over using BaseStationTimer, and it makes
debugging in simulation impossible. It now uses BaseStationTimer to
drive internal state transitions.

(cherry picked from commit 83a8eee9e18b76e9360185d9ac1a1ab86e8b97ed)

+++ engine/aiComponent/behaviorComponent/behaviors/timer/behaviorAdvanceClock.cpp
@@ -105,7 +105,7 @@ void BehaviorAdvanceClock::TransitionToShowClockInternal()
     audioMessage.audioEvent = AudioMetaData::GameEvent::GenericEvent::Stop__Robot_Vic_Sfx__Timer_Run_Down_Loop_Stop;
 
     RobotInterface::EngineToRobot wrapper(std::move(audioMessage));
-    GetBEI().GetAnimationComponent().AlterStreamingAnimationAtTime(std::move(wrapper), Util::SecToMilliSec(static_cast<float>(GetTimeDisplayClock_sec())));
+    GetBEI().GetAnimationComponent().AlterStreamingAnimationAtTime(std::move(wrapper), Util::SecToMilliSec(GetTimeDisplayClock_sec()));
   }
 }
